Tests Services and part of Controllers

+++ src/main/java/com/invoiceservice/invoiceservice/business/repository/model/OrderDetailsDAO.java
@@ -38,7 +38,7 @@ public class OrderDetailsDAO {
 
     @Column(name = "quantity")
     @Positive
-    private Float quantity;
+    private int quantity;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "invoice_id")
